controller and test added
